♻️ refactor: #54 MarkerContainer 컴포넌트 분리
